Move string operations to a separate folder
There will be more of these - regexes are getting too slow as the
complexity of the diff increases.
