+++ src/string/addStringToDiff.ts
